Refactor project structure and add new features
- Added `.gitignore` to exclude unnecessary files
- Moved and renamed CSS files to `public/css/` for better organization
- Added `cart-icon.png` and upload functionality in `public/uploads/`
- Deleted unused `user_dashboard.css` and `partials/navbar.ejs`
- Refactored project structure:
  - Added `src/` folder with routes, middleware, config, and utils
  - Separated routes into individual files: `admin.js`, `auth.js`, `developer.js`, `game.js`, `payment.js`, `user.js`
  - Created `src/utils/fileUpload.js` for handling file uploads
- Improved views:
  - Added `partials/header.ejs` and `partials/footer.ejs` for layout consistency
  - Modified EJS templates for cart, dashboard, game details, home, and payment pages
- Updated `server.js` to integrate new routes and project structure
- Updated `package.json` and `package-lock.json`
- Modified database file `game_store.db`

This commit improves project modularity, enhances layout consistency, and organizes assets and source code effectively.
