+++ package.json
@@ -14,6 +14,7 @@
     "express": "^4.21.2",
     "express-session": "^1.18.1",
     "multer": "^1.4.5-lts.1",
+    "projectwebapp": "file:",
     "sqlite3": "^5.1.7"
   }
 }
