+++ public/css/register.css
@@ -1,0 +1,106 @@
+/* General Body Style */
+body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    background-color: #1b2838; /* Epic Gamer dark theme */
+    color: #c7d5e0; /* Light text */
+    height: 100vh; /* Full viewport height for centering */
+    display: flex; /* Flexbox for centering */
+    justify-content: center; /* Center horizontally */
+    align-items: center; /* Center vertically */
+}
+
+/* Form Container Style */
+.form-container {
+    background-color: #2a475e; /* Darker background for contrast */
+    border-radius: 10px;
+    box-shadow: 0 0 15px rgba(0, 0, 0, 0.5); /* Subtle shadow for depth */
+    padding: 30px;
+    max-width: 400px; /* Increased width for better appearance */
+    width: 100%; /* Ensure responsiveness */
+    text-align: center;
+}
+
+/* Header Style */
+h2 {
+    font-size: 28px;
+    color: #66c0f4; /* Accent color */
+    margin-bottom: 20px; /* Space below header */
+}
+
+/* Form Style */
+form {
+    display: flex;
+    flex-direction: column; /* Stack inputs vertically */
+    align-items: center; /* Center align items */
+}
+
+form select,
+form input {
+    width: 90%; /* Full width with padding */
+    padding: 10px;
+    margin: 10px 0;
+    border: none;
+    border-radius: 5px;
+    background-color: #1b2838;
+    color: #c7d5e0;
+    font-size: 16px;
+}
+
+/* Show/hide Studio Name Field */
+form #studioName {
+    display: none;
+}
+
+form button {
+    width: 90%;
+    padding: 10px;
+    background-color: #66c0f4;
+    border: none;
+    border-radius: 5px;
+    color: white;
+    font-size: 18px;
+    cursor: pointer; /* Pointer cursor for button */
+}
+
+form button:hover {
+    background-color: #417a9b;
+}
+
+/* Notification Styles */
+.notification {
+    display: none; /* Hidden by default */
+    margin-bottom: 10px;
+    padding: 10px;
+    border-radius: 5px;
+    font-size: 14px;
+    text-align: center;
+}
+
+.notification.error {
+    background-color: #f8d7da;
+    color: #842029;
+    border: 1px solid #f5c2c7;
+}
+
+.notification.success {
+    background-color: #d1e7dd;
+    color: #0f5132;
+    border: 1px solid #badbcc;
+}
+
+/* Footer */
+p {
+    margin-top: 20px;
+    color: #c7d5e0;
+}
+
+p a {
+    color: #66c0f4;
+    text-decoration: none;
+}
+
+p a:hover {
+    color: #417a9b;
+}
