--- public/login.css
@@ -1,125 +1,0 @@
-/* General Body Style for Login Page */
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #1b2838; /* Epic Gamer dark theme */
-    color: #c7d5e0; /* Light text */
-    height: 100vh; /* Full viewport height */
-    display: flex;
-    justify-content: center; /* Center horizontally */
-    align-items: center; /* Center vertically */
-}
-
-/* Login Form Container */
-.form-container { /* Changed from .login-container for consistency with the register page */
-    background-color: #2a475e; /* Darker background for contrast */
-    border-radius: 10px;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.5); /* Shadow for depth */
-    padding: 30px; /* Adjusted padding to match register page */
-    max-width: 400px; /* Consistent width with register page */
-    width: 100%; /* Ensure responsiveness */
-    text-align: center;
-}
-
-/* Title (Login header) */
-h2 {
-    font-size: 28px; /* Adjusted font size to match register page */
-    color: #66c0f4; /* Accent color */
-    margin-bottom: 20px;
-}
-
-/* Form Layout */
-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-/* Input Fields */
-form input {
-    width: 90%; /* Full width with padding */
-    padding: 10px;
-    margin: 10px 0;
-    border: none;
-    border-radius: 5px;
-    background-color: #1b2838;
-    color: #c7d5e0;
-    font-size: 16px;
-}
-
-/* Input Focus Effect */
-form input:focus {
-    outline: none;
-    border: 1px solid #66c0f4;
-}
-
-/* Submit Button */
-form button {
-    width: 90%;
-    padding: 10px;
-    background-color: #66c0f4;
-    border: none;
-    border-radius: 5px;
-    color: white;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-/* Button Hover Effect */
-form button:hover {
-    background-color: #417a9b;
-}
-
-/* Footer Links for Login Page */
-p {
-    margin-top: 20px;
-    font-size: 14px;
-    color: #c7d5e0;
-}
-
-p a {
-    color: #66c0f4;
-    text-decoration: none;
-}
-
-p a:hover {
-    color: #417a9b;
-}
-
-/* Notification Styles */
-.notification {
-    display: none; /* Hidden by default */
-    margin-bottom: 10px;
-    padding: 10px;
-    border-radius: 5px;
-    font-size: 14px;
-    text-align: center;
-}
-
-.notification.error {
-    background-color: #f8d7da;
-    color: #842029;
-    border: 1px solid #f5c2c7;
-}
-
-.notification.success {
-    background-color: #d1e7dd;
-    color: #0f5132;
-    border: 1px solid #badbcc;
-}
-
-/* Add animations to notifications */
-.notification {
-    animation: fadeIn 0.5s ease-in-out;
-}
-
-@keyframes fadeIn {
-    0% {
-        opacity: 0;
-    }
-    100% {
-        opacity: 1;
-    }
-}
