+++ src/config/db.js
@@ -1,0 +1,10 @@
+const sqlite3 = require('sqlite3').verbose();
+const db = new sqlite3.Database('./game_store.db', (err) => {
+    if (err) {
+        console.error('Database connection error:', err.message);
+    } else {
+        console.log('Connected to the SQLite database.');
+    }
+});
+
+module.exports = db;
