+++ views/home.ejs
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Epic Gamer - Home</title>
-    <link rel="stylesheet" href="/styles.css">
+    <title>Steam-wannabe - Home</title>
+    <link rel="stylesheet" href="/css/styles.css">
 </head>
 <body>
     <header>
