+++ views/login_page.ejs
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Login</title>
-  <link rel="stylesheet" href="/login.css">
+  <link rel="stylesheet" href="/css/login.css">
 </head>
 <body>
   <div class="form-container">
