+++ views/partials/footer.ejs
@@ -1,0 +1,56 @@
+<!-- Footer -->
+<footer class="bg-dark text-light py-3 mt-auto">
+    <div class="container">
+        <div class="row text-center text-md-left">
+            <!-- About section -->
+            <div class="col-md-4 mb-3">
+                <h6 class="text-uppercase font-weight-bold">About Us</h6>
+                <p class="small mb-0">
+                    Steam-wannabe is your one-stop platform for exploring, purchasing, and enjoying the best games.
+                </p>
+            </div>
+            <!-- Quick Links -->
+            <div class="col-md-4 mb-3">
+                <h6 class="text-uppercase font-weight-bold">Quick Links</h6>
+                <ul class="list-unstyled small mb-0">
+                    <li><a class="text-light text-decoration-none" href="/dashboard">Home</a></li>
+                    <li><a class="text-light text-decoration-none" href="/dashboard">Store</a></li>
+                    <li><a class="text-light text-decoration-none" href="#">Support</a></li>
+                    <li><a class="text-light text-decoration-none" href="/logout">Logout</a></li>
+                </ul>
+            </div>
+            <!-- Contact Us -->
+            <div class="col-md-4 mb-3">
+                <h6 class="text-uppercase font-weight-bold">Contact Us</h6>
+                <ul class="list-unstyled small mb-0">
+                    <li>Email: <a class="text-light" href="mailto:support@steam-wannabe.com">support@steam-wannabe.com</a></li>
+                    <li>Phone: +1 (234) 567-890</li>
+                    <li>Address: 123 Game Street, Virtual City</li>
+                </ul>
+            </div>
+        </div>
+        <hr class="bg-secondary my-3">
+        <div class="text-center small">
+            &copy; 2024 Steam-wannabe. All Rights Reserved.
+        </div>
+    </div>
+</footer>
+
+<style>
+    html, body {
+        height: 100%;
+        margin: 0;
+    }
+    body {
+        display: flex;
+        flex-direction: column;
+        min-height: 100vh;
+    }
+    .content {
+        flex: 1;
+    }
+    footer {
+        padding-top: 20px;
+        padding-bottom: 20px;
+    }
+</style>
