--- views/partials/navbar.ejs
@@ -1,24 +1,0 @@
-<!-- Navbar -->
-<nav class="navbar navbar-expand-md navbar-dark bg-dark">
-    <a class="navbar-brand" href="/dashboard">Steam</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav">
-        <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-            <li class="nav-item"><a class="nav-link" href="/dashboard">Home</a></li>
-            <li class="nav-item"><a class="nav-link" href="/dashboard">Store</a></li>
-            <li class="nav-item"><a class="nav-link" href="#">Support</a></li>
-        </ul>
-        <!-- Cart button on right top -->
-        <ul class="navbar-nav ml-auto">
-            <li class="nav-item">
-                <a class="nav-link" href="/cart">
-                    <img src="/images/cart-icon.png" alt="Cart" style="height: 24px; width: 24px;">
-                    Cart
-                </a>
-            </li>
-            <li class="nav-item"><a class="nav-link" href="/logout">Logout</a></li>
-        </ul>
-    </div>
-</nav>
