+++ views/payment_page.ejs
@@ -61,7 +61,7 @@
 
             if (response.ok) {
                 notification.className = 'alert alert-success';
-                notification.textContent = text || 'Payment successful! Games have been added to your library, You gonna be redirected to the homepage in 3s.';
+                notification.textContent = text || 'Payment successful! Games have been added to your library, you gonna be redirected to the homepage in 3s.';
                 notification.classList.remove('d-none');
 
                 // Chuyển về trang chính sau 3 giây
