+++ views/register_page.ejs
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Register</title>
-  <link rel="stylesheet" href="register.css">
+  <link rel="stylesheet" href="/css/register.css">
   <style>
   /* Add notification styles */
   .notification {
