+++ views/user_dashboard.ejs
@@ -15,66 +15,73 @@
 </head>
 <body style="background-color: #1b2838; color: white;">
 
-<!-- Include Global Navbar -->
-<%- include('partials/navbar') %>
+    <!-- Include Global Navbar -->
+    <%- include('partials/header') %>
     
-    <div class="container mt-4">
-        <h1>Welcome, <%= user.username %>!</h1>
-        <p>Explore our collection of games below!</p>
+    <!-- Main Content -->
+    <div class="content">
+        <div class="container mt-4">
+            <h1>Welcome, <%= user.username %>!</h1>
+            <p>Explore our collection of games below!</p>
 
-    <!-- Search and Filter Form -->
-    <form class="form-inline mb-3" method="GET" action="/dashboard">
-        <!-- Search -->
-        <input type="text" name="search" class="form-control mr-2" placeholder="Search by title" value="<%= search || '' %>">
+            <!-- Search and Filter Form -->
+            <form class="form-inline mb-3" method="GET" action="/dashboard">
+                <!-- Search -->
+                <input type="text" name="search" class="form-control mr-2" placeholder="Search by title" value="<%= search || '' %>">
 
-        <!-- Filter by Category -->
-        <select name="category" class="form-control mr-2">
-            <option value="">All Categories</option>
-            <option value="Action" <%= category === 'Action' ? 'selected' : '' %>>Action</option>
-            <option value="Adventure" <%= category === 'Adventure' ? 'selected' : '' %>>Adventure</option>
-            <option value="Strategy" <%= category === 'Strategy' ? 'selected' : '' %>>Strategy</option>
-            <option value="Sports" <%= category === 'Sports' ? 'selected' : '' %>>Sports</option>
-            <option value="Role-Playing Game - RPG" <%= category === 'Role-Playing Game - RPG' ? 'selected' : '' %>>Role-Playing Game - RPG</option>
-            <option value="First-Person Shooter - FPS" <%= category === 'First-Person Shooter - FPS' ? 'selected' : '' %>>First-Person Shooter - FPS</option>
-            <option value="Multiplayer Online Battle Arena - MOBA" <%= category === 'Multiplayer Online Battle Arena - MOBA' ? 'selected' : '' %>>Multiplayer Online Battle Arena - MOBA</option>
-            <option value="Survival" <%= category === 'Survival' ? 'selected' : '' %>>Survival</option>
-            <option value="Simulation" <%= category === 'Simulation' ? 'selected' : '' %>>Simulation</option>
-            <option value="Educational" <%= category === 'Educational' ? 'selected' : '' %>>Educational</option>
-        </select>
+                <!-- Filter by Category -->
+                <select name="category" class="form-control mr-2">
+                    <option value="">All Categories</option>
+                    <option value="Action" <%= category === 'Action' ? 'selected' : '' %>>Action</option>
+                    <option value="Adventure" <%= category === 'Adventure' ? 'selected' : '' %>>Adventure</option>
+                    <option value="Strategy" <%= category === 'Strategy' ? 'selected' : '' %>>Strategy</option>
+                    <option value="Sports" <%= category === 'Sports' ? 'selected' : '' %>>Sports</option>
+                    <option value="Role-Playing Game - RPG" <%= category === 'Role-Playing Game - RPG' ? 'selected' : '' %>>Role-Playing Game - RPG</option>
+                    <option value="First-Person Shooter - FPS" <%= category === 'First-Person Shooter - FPS' ? 'selected' : '' %>>First-Person Shooter - FPS</option>
+                    <option value="Multiplayer Online Battle Arena - MOBA" <%= category === 'Multiplayer Online Battle Arena - MOBA' ? 'selected' : '' %>>Multiplayer Online Battle Arena - MOBA</option>
+                    <option value="Survival" <%= category === 'Survival' ? 'selected' : '' %>>Survival</option>
+                    <option value="Simulation" <%= category === 'Simulation' ? 'selected' : '' %>>Simulation</option>
+                    <option value="Educational" <%= category === 'Educational' ? 'selected' : '' %>>Educational</option>
+                </select>
 
-        <!-- Submit Button -->
-        <button type="submit" class="btn btn-primary">Search</button>
-    </form>
+                <!-- Submit Button -->
+                <button type="submit" class="btn btn-primary">Search</button>
+            </form>
 
-    <!-- Game Grid -->
-    <div class="row">
-        <% if (games.length === 0) { %>
-            <!-- Display this message if no games are found -->
-            <div class="col-12 text-center">
-                <p class="text-muted">No games found for your search criteria.</p>
-            </div>
-        <% } else { %>
-            <!-- Render games if found -->
-            <% games.forEach(game => { %>
-                <div class="col-lg-3 col-md-4 col-sm-6 mb-4">
-                    <!-- Wrap the card in an anchor tag -->
-                    <a href="/game/<%= game.GameID %>" class="text-decoration-none">
-                        <div class="card bg-dark text-white">
-                            <img src="<%= game.Image %>" class="card-img-top" alt="<%= game.Title %>">
-                            <div class="card-body">
-                                <h5 class="card-title"><%= game.Title %></h5>
-                                <p class="card-text">$<%= game.Price.toFixed(2) %></p>
-                                <p class="card-text"><%= game.Description %></p>
-                                <p class="card-text">
-                                    <strong>Category:</strong> <%= game.Category %>
-                                </p>
-                            </div>
+            <!-- Game Grid -->
+            <div class="row">
+                <% if (games.length === 0) { %>
+                    <!-- Display this message if no games are found -->
+                    <div class="col-12 text-center">
+                        <p class="text-muted">No games found for your search criteria.</p>
+                    </div>
+                <% } else { %>
+                    <!-- Render games if found -->
+                    <% games.forEach(game => { %>
+                        <div class="col-lg-3 col-md-4 col-sm-6 mb-4">
+                            <a href="/game/<%= game.GameID %>" class="text-decoration-none">
+                                <div class="card bg-dark text-white">
+                                    <img src="<%= game.Image %>" class="card-img-top" alt="<%= game.Title %>">
+                                    <div class="card-body">
+                                        <h5 class="card-title"><%= game.Title %></h5>
+                                        <p class="card-text">$<%= game.Price.toFixed(2) %></p>
+                                        <p class="card-text"><%= game.Description %></p>
+                                        <p class="card-text">
+                                            <strong>Category:</strong> <%= game.Category %>
+                                        </p>
+                                    </div>
+                                </div>
+                            </a>
                         </div>
-                    </a>
-                </div>
-            <% }) %>
-        <% } %>
+                    <% }) %>
+                <% } %>
+            </div>
+        </div>
     </div>
-</div>
+
+    <!-- Include Global Footer -->
+    <%- include('partials/footer') %>
+
 </body>
+
 </html>
